backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
aiofiles==24.1.0
httpx==0.28.1
beautifulsoup4==4.12.3
playwright==1.49.1
pydantic==2.10.4
pydantic-settings==2.7.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
rembg==2.0.59
onnxruntime==1.20.1
scikit-image==0.24.0
